requirements.txt
ipython~=6.29.3
numpy~=1.26.4
scamp~=0.9.2.post2 # also sudo apt install fluidsynth
PyYAML~=6.0.1